ingestion/requirements.txt
openai>=1.109.1
supabase==2.11.0
python-dotenv==1.0.1